Api. Move shared data classes
