+++ src/back/Storage/Clone/TopicsUntracked.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace KeepersTeam\Webtlo\Storage\Clone;
 
-use KeepersTeam\Webtlo\External\Api\V1\TopicDetails;
+use KeepersTeam\Webtlo\External\Data\TopicDetails;
 use KeepersTeam\Webtlo\Storage\CloneTable;
 use KeepersTeam\Webtlo\Storage\Table\Topics;
 use Psr\Log\LoggerInterface;
