+++ src/back/Update/TorrentsClients.php
@@ -9,10 +9,10 @@ use KeepersTeam\Webtlo\Config\SubForums;
 use KeepersTeam\Webtlo\Config\TopicSearch;
 use KeepersTeam\Webtlo\Config\TorrentClients;
 use KeepersTeam\Webtlo\Enum\UpdateMark;
-use KeepersTeam\Webtlo\External\Api\V1\TopicDetails;
-use KeepersTeam\Webtlo\External\Api\V1\TopicSearchMode;
 use KeepersTeam\Webtlo\External\ApiReportClient;
 use KeepersTeam\Webtlo\External\Data\ApiError;
+use KeepersTeam\Webtlo\External\Data\TopicDetails;
+use KeepersTeam\Webtlo\External\Data\TopicSearchMode;
 use KeepersTeam\Webtlo\External\ForumClient;
 use KeepersTeam\Webtlo\Storage\Clone\TopicsUnregistered;
 use KeepersTeam\Webtlo\Storage\Clone\TopicsUntracked;
